Extract utility test code to a separate fixture
